requirements.txt
streamlit==1.20.0
requests==2.28.1
beautifulsoup4==4.11.1
python-whois==0.9.1